Initialize support for PayPal types.
Following the existing PayPal SDKs in other languages the Python
implementation will support generating API payloads in an OOP manner.

This is the initial commit of the PayPal types in which base classes
are implemented. Allowing us to write the PayPal types in an elegant
manner much like fields in an ORM; supporting sanitization and
validation of given values.

The design choice to validate and sanitize field values is due to two
reasons. One being to avoid the generation of API requests which will
undoubtly result in a failure. The other is to document the API through
code rather than having to deal with the PayPal documentation
(hopefully).