Adding scrape_method_documentation.py script.
This script is capable of fetching the documentation corresponding to
the given API service and method and then converting it to our intended
PayPal type docstrings.

Thereby saving manual labour in writing these docstrings along with the
possibility to easily detect future changes in the documentation and
thus the API itself.