More robust retrieval of types group name.
Some API methods do not conform to the convention of specifying the
group name within the h4.topictitle3 tag. However, they still specify it
within the h4 tag, but without the topictitle3 class. By removing the
explicit requirement of the topictitle3 class we can retrieve the group
name for all API methods.